+++ sentry-rails/examples/rails-6.0/app/jobs/error_job.rb
@@ -1,5 +1,5 @@
 class ErrorJob < ApplicationJob
   def perform
-    1 / 0
+    raise "Job failed"
   end
 end